pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64", "setuptools-scm>=8"]

[project]
name = "nat_sfc_router"
dynamic = ["version"]
dependencies = [
    "nvidia-nat[phoenix]",
    "torch",
    "transformers==4.51.3",
    "numpy<2.00",
    "scikit-learn",
    "openai",
    "requests",
    "clip-client",
    "joblib",
    "tqdm",
    "Pillow",
    "datasets",
    "gradio",
    "python-dotenv",
    "docarray==0.21.0"
]
requires-python = ">=3.11.0"
description = "Simple Calculator NeMo Agent toolkit example"
keywords = ["ai", "rag", "agents"]
classifiers = ["Programming Language :: Python"]

[project.entry-points.'nat.components']
nat_multimodal_agent = "nat_sfc_router.register"

[tool.setuptools.package-data]
nat_sfc_router = [
    "*.npy",
    "*.json",
    "*.pt",
    "*.pth",
    "*.yml",
    "configs/*.yml",
    "training/router_artifacts/*"
]
